friends component with user card created
