--- src/components/Friends.js
@@ -1,31 +1,0 @@
-const friends = [
-	{
-		name: 'Leonard Krasner',
-		handle: 'leonardkrasner',
-		imageUrl:
-			'https://images.unsplash.com/photo-1519345182560-3f2917c472ef?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-		href: '#',
-	},
-	{
-		name: 'Floyd Miles',
-		handle: 'floydmiles',
-		imageUrl:
-			'https://images.unsplash.com/photo-1463453091185-61582044d556?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-		href: '#',
-	},
-	{
-		name: 'Emily Selman',
-		handle: 'emilyselman',
-		imageUrl:
-			'https://images.unsplash.com/photo-1502685104226-ee32379fefbe?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-		href: '#',
-	},
-	{
-		name: 'Kristin Watson',
-		handle: 'kristinwatson',
-		imageUrl:
-			'https://images.unsplash.com/photo-1500917293891-ef795e70e1f6?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-		href: '#',
-	},
-];
-export default friends;
